+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Added -a option to allow for append of data to an existing output file.
 
 ### Changed
+- status:  Added file mode option to writing data to a file.  Default is write.
 - config/mongo.py.TEMPLATE:  Changed to generic setup.
 - config/mysql.cfg.TEMPLATE:  Changed to generic setup.
 - config/mysql_cfg.py.TEMPLATE:  Changed to generic setup.